+++ run.sh
@@ -1,3 +1,8 @@
+if [ ! -e nginx/proxy_prod.conf ]; then
+  echo "Create nginx/proxy_prod.conf file to configure reverse proxy on production environment."
+  exit 1
+fi
+
 cp nginx/proxy_prod.conf nginx/proxy.conf
 
 docker-compose down
